combined question and answer api data into one resulting object...
the resulting object is a QuestionMk2

QuestionMk2 consists of a question ID, question text, and an arraylist of answers. Each element of the answer arraylist consists of a answer id, answer text, and a boolean IsCorrect field. 

The next task is to attempt to rework the QuizQuestionsActivity to accommodate this change